seperated middleware utils into seperate utils.py file
